refactor push model client
